requirements.txt
# Development Dependencies
setuptools>=65.5.0
wheel>=0.38.0

# API and Data Handling
ccxt==3.1.54
pandas==2.0.3
pandas-ta==0.3.14b0
numpy==1.24.3

# Web Framework
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0

# Database
sqlalchemy==2.0.20
alembic==1.12.0

# Notifications
python-telegram-bot==20.5

# Utilities
python-dotenv==1.0.0
schedule==1.2.0
backoff==2.2.1

# Visualization (optional)
plotly==5.16.1
dash==2.15.0

# Testing
pytest==7.4.2